Rename `unify` modules to `dispatch` and clean up Theano tests
