+++ Makefile
@@ -64,10 +64,8 @@ coverage: test
 	diff-cover coverage.xml --compare-branch=master --fail-under=100
 
 docs:
-	# latesttag=$(git describe --tags `git rev-list --tags --max-count=1`)
-	# echo checking out ${latesttag}
-	# git checkout ${latesttag}
 	pushd docs && \
+	sphinx-apidoc ../symbolic_pymc/ -o source --force && \
 	make html && \
 	ghp-import -n -p -r upstream -b gh-pages build/html && \
 	popd
