+++ docs/source/symbolic_pymc.rst
@@ -13,18 +13,18 @@ Subpackages
 Submodules
 ----------
 
-symbolic\_pymc.meta module
---------------------------
+symbolic\_pymc.dispatch module
+------------------------------
 
-.. automodule:: symbolic_pymc.meta
+.. automodule:: symbolic_pymc.dispatch
    :members:
    :undoc-members:
    :show-inheritance:
 
-symbolic\_pymc.unify module
----------------------------
+symbolic\_pymc.meta module
+--------------------------
 
-.. automodule:: symbolic_pymc.unify
+.. automodule:: symbolic_pymc.meta
    :members:
    :undoc-members:
    :show-inheritance:
