+++ docs/source/symbolic_pymc.tensorflow.rst
@@ -4,6 +4,22 @@ symbolic\_pymc.tensorflow package
 Submodules
 ----------
 
+symbolic\_pymc.tensorflow.dispatch module
+-----------------------------------------
+
+.. automodule:: symbolic_pymc.tensorflow.dispatch
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+symbolic\_pymc.tensorflow.graph module
+--------------------------------------
+
+.. automodule:: symbolic_pymc.tensorflow.graph
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 symbolic\_pymc.tensorflow.meta module
 -------------------------------------
 
@@ -16,14 +32,6 @@ symbolic\_pymc.tensorflow.printing module
 -----------------------------------------
 
 .. automodule:: symbolic_pymc.tensorflow.printing
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-symbolic\_pymc.tensorflow.unify module
---------------------------------------
-
-.. automodule:: symbolic_pymc.tensorflow.unify
    :members:
    :undoc-members:
    :show-inheritance:
