+++ symbolic_pymc/tensorflow/__init__.py
@@ -1,2 +1,2 @@
 # Needed to register generic functions
-from .unify import *
+from .dispatch import *
